props, filters are used
db.json is created to store data, I opened the JSON server in a separate port number. But still, the main react frontend shows data from the data array in Home.js. Now I will proceed to change this system, so committing and saving up to this implementation in a separate branch.